logthrdestdrv: do an initial flush at startup & a final flush at exit
When the threaded destination exits, the batch may contain pending messages
(due to flush-timeout() or partial flushes).  This patch adds an explicit
flush at exit time, and rewinds the remainder of the batch, should it still
exist.

It also adds an initial flush, so should the queue still contain elements on
its backlog at startup (this can only happen with a disk buffer), we'd
rewind and resend them.

This was discussed here:

https://github.com/balabit/syslog-ng/pull/2098#issuecomment-421097329

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>