xml-scanner: make it inheritable
Usually we allocate the scanners during message passing, but to save
memory allocation's overhead we store them on the stack.

To be able to override the callbacks used by g_markup, init() function
will only setup the virtual functions.
The g_markup context will be created in the parse() function and only exists
while the parsing lasts.
Hence, we don't need GMarkup as an XMLScanner class member.
Of course, this can look a bit ugly in a UML class diagram as g_markup is hidden.