xml-scanner: remove GMarkupContext from virtual functions
If removed, GMarkupContext can be a private member.

Also, if we split the markup callbacks into 2 separate functions we can add a condition when we want to push.
(one has every functionality except the actual pushing, the other calls the first one and does the key-value pushing)
This can be handy in subclasses where we would like to override the parent class' logic of when should it push a name-value pair.

This patch had to re-introduce the GMarkupContext as an XMLScanner class member to be able to use it internally when needed.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>