xml: add text-separator() option
XMl-parser's current limitation is that it cannot handle array-like
structures in the XML.

Example
```
<List>
  <Data>Info 1</Data>
  <Data>Info 2</Data>
</List>
```

In this case we just concatenate a different tag's content:
List.Data = "Info 1Info2"

To be able to separate the different contents the xml-parser
has a new option, called text-separator() which will be added
each time an existing tag is updated.
By default, the '\n' character is used.

Result
List.Data = "Info 1\nInfo2"

IMPORTANT
xml tags with mixed content (text and child element) should not have their text content
concatenated with the text-separator, although this is broken currently

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>