xml-scanner: Buffer partial text contents inside xml-scanner
Partial text contents now buffered inside xml-scanner
until the end tag is not found.
This way we can solve the text-separator option problem on parser side,
which should be used only in case of array-like structures and not on
mixed contents.

The buffering implemented by stacking the current text content, and
whenever a start/end tag found we push/pop the current text content.
The push-key-value() call moved from text() to end() method so we don't
push chunks.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>