lib/cfg-grammar: use api.prefix instead of name-prefix
This patch is intended to fix bison 3.3 deprecation warnings, which started
in bison 3.3, which deprecated the use of name-prefix, with a preference
towards the api.prefix define.

Unfortunately api.prefix didn't fully support our use-case in its original
form, so we require at least bison 3.4.2 that already includes the
api.location.type feature for the C language.

https://github.com/akimd/bison/commit/9260e5ca4fa62ff971fc52a2dd64474dd60e5990

By merging this branch to master, we would start depending on bison
3.4.2.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>