cfg-lexer: establish YYSTYPE/YYLTYPE naming conventions
YYLTYPE/YYSTYPE naming conventions

We use multiple bison generated grammars (basically one for each plugin)
with a single lexer and the same location/symbol types.  Earlier we could
easily just define YYLTYPE and YYSTYPE here and all generated grammars
and the lexer used it properly.  With the advent of the `api.prefix'
option for the grammars (and the deprecation of the old `name-prefix'
behaviors), we needed to complicate things somewhat.

We have three contexts where we need to use proper type names:
  - in our own code where we might need to use location information (e.g. YYLTYPE)
  - in the generated lexer,
  - in the generated grammars

Our own code
============
Because of the various #define/typedef games that generated code uses, I
decided that our own code should not use the names YYLTYPE/YYSTYPE
directly.  In those cases we use CFG_LTYPE and CFG_STYPE to indicate that
these are types related to our configuration language.  None of the
grammars use the "CFG_" prefix (and should not in the future either).

The generated lexer
===================
The lexer get these types by us #define-ing YYLTYPE/YYSTYPE to
CFG_LTYPE/STYPE but only privately, e.g.  these definitions should not be
published to the rest of the codebase.  We do this by defining these in
implementation files and not in the headers.  This is because some of the
code would try to #ifdef based on the existance of these macros.

The generated grammars
======================
The grammars each have an api.location.type and api.value.type options in
their .y files, which use the names CFG_LTYPE and CFG_STYPE respectively.
The generated code uses YYLTYPE and YYSTYPE internally (defined as
macros), but because of the previous points this does not create a
conflict.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>